feat: add ability to run unleash
Added unleash to the docker compose by copying theirs from github, along with the scripts directory
which they use for health checks and apparently some other stuff. Got the server running and could
log in and create a feature flag, and create an API key. Next step is to hook up the app.